+++ example/Interactive Figma Desktop Prototype/src/app/App.tsx
@@ -1,9 +1,0 @@
-import { GnomeDesktop } from "./components/GnomeDesktop";
-
-export default function App() {
-  return (
-    <div className="w-screen h-screen overflow-hidden relative">
-      <GnomeDesktop />
-    </div>
-  );
-}