+++ src/constants/other.ts
@@ -10,10 +10,6 @@ export const COMMUNITY_TAB_URL = `${HOMEPAGE}/@figma_linux`;
 // export const PARSED_HOMEPAGE = parse("https://www.figma.com");
 export const PARSED_HOMEPAGE = new URL("https://www.figma.com");
 
-export const DOWNLOAD_ZIP_URI =
-  "https://github.com/Figma-Linux/figma-linux-themes/archive/master.zip";
-export const DOWNLOAD_ZIP_PATH = "/tmp/master.zip";
-
 export const TOPPANELHEIGHT = 40;
 export const MENU_WIDTH = 330;
 
@@ -21,13 +17,11 @@ export const STARTUP_DELAY_MS = 1500;
 
 export const PROTOCOL = "figma";
 
-export const TEST_THEME_ID = "test-creator-theme";
-
 export const MANIFEST_FILE_NAME = "manifest.json";
 export const CHROME_GPU = "chrome://gpu";
 
-export const CONFIGDIR = `${_env.HOME || ""}/.config/figma-linux`;
-export const RESOURCESDIR = `${_env.HOME || ""}/.config/figma-linux/resources`;
+export const CONFIGDIR = `${_env.HOME || ""}/.config/figma-linux-next`;
+export const RESOURCESDIR = `${_env.HOME || ""}/.config/figma-linux-next/resources`;
 export const REGEXP_APP_AUTH_GRANT = /^\/{0,2}app_auth\/[^/]+\/grant/;
 
 export const FIGMA_SESSION_COOKIE_NAME = "figma.session";
@@ -49,11 +43,11 @@ export const NEW_FILE_TAB_TITLE = "New file";
 export const LINKS = {
   HELP_PAGE: "https://help.figma.com",
   PLUGINS_DOCS: `${HOMEPAGE}/plugin-docs/intro`,
-  FIGMA_COMMUNITY_FORUM: "https://spectrum.chat/figma",
-  FIGMA_LINUX_COMMUNITY_FORUM: "https://spectrum.chat/figma-linux",
+  FIGMA_COMMUNITY_FORUM: "https://github.com/arximus88/figma-linux-next/discussions",
+  FIGMA_LINUX_COMMUNITY_FORUM: "https://github.com/arximus88/figma-linux-next/discussions",
   FIGMA_LINUX_TELEGRAM: "https://t.me/figma_linux",
-  VIDEO_TUTORIALS: "https://www.youtube.com/figmadesign",
-  RELEASE_NOTES: "https://github.com/Figma-Linux/figma-linux/releases/latest",
-  LEGAL_SUMMARY: `${HOMEPAGE}/summary-of-policy`,
   THEMES_REPO: "https://github.com/Figma-Linux/figma-linux-themes",
+  VIDEO_TUTORIALS: "https://www.youtube.com/figmadesign",
+  RELEASE_NOTES: "https://github.com/arximus88/figma-linux-next/releases/latest",
+  LEGAL_SUMMARY: `${HOMEPAGE}/summary-of-policy`,
 };