+++ src/main/controllers/SettingsController.ts
@@ -39,9 +39,10 @@ export default class SettingsController {
       (event: IpcMainInvokeEvent) => event.sender.isDevToolsOpened(),
       "SettingsController",
     );
+    ipcRegistry.on("setFrameStyle", this.setFrameStyle.bind(this), "SettingsController");
   }
 
-  private closeSettingsView(_: IpcMainEvent, settings: Types.SettingsInterface) {
+  private async closeSettingsView(_: IpcMainEvent, settings: Types.SettingsInterface) {
     if (storage.settings.app.enableColorSpaceSrgb !== settings.app.enableColorSpaceSrgb) {
       app.emit("enableColorSpaceSrgbWasChanged", true);
     }
@@ -56,7 +57,7 @@ export default class SettingsController {
     }
 
     storage.settings = settings;
-    storage.save();
+    await storage.save();
 
     this.windowManager.closeSettingsViewForLastWindow();
   }
@@ -77,5 +78,11 @@ export default class SettingsController {
 
   private updateFigmaUiScale(_: IpcMainInvokeEvent, scale: number) {
     this.windowManager.updateFigmaUiScaleAllWindows(scale);
+  }
+
+  private setFrameStyle(_: IpcMainEvent, style: Types.FrameStyle) {
+    storage.settings.app.frameStyle = style;
+    storage.save();
+    this.windowManager.setFrameStyleAllWindows(style);
   }
 }