+++ src/renderer/Common/Store/Themes/index.ts
@@ -1,1 +1,0 @@
-export { themeApp } from "./themeApp.svelte";